+++ ui/components-react/src/components-react/editors/NumericInputEditor.tsx
@@ -32,6 +32,7 @@ import { InputWithDecorations } from "@itwin/itwinui-react";
 interface NumericInputEditorState {
   value: number;
   size?: number;
+  maxSize?: number;
   maxLength?: number;
 
   min?: number;
@@ -129,6 +130,7 @@ export class NumericInputEditor
     }
 
     let size: number | undefined;
+    let maxSize: number | undefined;
     let maxLength: number | undefined;
     let min: number | undefined;
     let max: number | undefined;
@@ -147,6 +149,7 @@ export class NumericInputEditor
       ) as InputEditorSizeParams;
       if (editorSizeParams) {
         if (editorSizeParams.size) size = editorSizeParams.size;
+        if (editorSizeParams.maxSize) maxSize = editorSizeParams.maxSize;
         if (editorSizeParams.maxLength) maxLength = editorSizeParams.maxLength;
       }
 
@@ -166,6 +169,7 @@ export class NumericInputEditor
       this.setState({
         value: initialValue,
         size,
+        maxSize,
         maxLength,
         min,
         max,
@@ -185,6 +189,9 @@ export class NumericInputEditor
     const style: React.CSSProperties = {
       ...this.props.style,
       minWidth: `${minSize * 0.75}em`,
+      maxWidth: this.state.maxSize
+        ? `${this.state.maxSize * 0.75}em`
+        : undefined,
     };
 
     return (