+++ ui/components-react/src/components-react/editors/TextareaEditor.tsx
@@ -37,6 +37,7 @@ type TextareaProps = React.ComponentPropsWithoutRef<typeof Textarea>;
 interface TextareaEditorState {
   inputValue: string;
   size?: number;
+  maxSize?: number;
   maxLength?: number;
   rows: number;
 }
@@ -121,6 +122,7 @@ export class TextareaEditor
     }
 
     let size: number | undefined;
+    let maxSize: number | undefined;
     let maxLength: number | undefined;
     let rows: number = DEFAULT_ROWS;
 
@@ -136,6 +138,7 @@ export class TextareaEditor
       ) as InputEditorSizeParams;
       if (editorSizeParams) {
         if (editorSizeParams.size) size = editorSizeParams.size;
+        if (editorSizeParams.maxSize) maxSize = editorSizeParams.maxSize;
         if (editorSizeParams.maxLength) maxLength = editorSizeParams.maxLength;
       }
 
@@ -152,6 +155,7 @@ export class TextareaEditor
       this.setState({
         inputValue: initialValue,
         size,
+        maxSize,
         maxLength,
         rows,
       });
@@ -185,6 +189,9 @@ export class TextareaEditor
     const style: React.CSSProperties = {
       ...this.props.style,
       minWidth: `${minSize * 0.75}em`,
+      maxWidth: this.state.maxSize
+        ? `${this.state.maxSize * 0.75}em`
+        : undefined,
     };
     const textareaProps: TextareaProps = {
       className: "components-textarea-editor-textarea",