+++ ui/components-react/CHANGELOG.md
@@ -1,5 +1,40 @@
 # Change Log - @itwin/components-react
 
+## 5.28.1
+
+### Patch Changes
+
+- 5976547: Changed `MergedPropertyValueRenderer` to render `--` instead of `*** varies ***`. If a merged property has a display value starting with `--` (e.g. `-- unit`), that value will be used instead of the default `--`. Text editor now also displays `--` for merged properties instead of showing an empty value.
+- Updated dependencies [2a864ed]
+  - @itwin/core-react@5.28.1
+
+## 5.28.0
+
+### Minor Changes
+
+- 3585ac3: Deprecated `ControlledTree` component and all related tree APIs in favor of the `Tree` component from `@stratakit/structures`.
+
+  Added `HIGHLIGHT_ACTIVE_CLASS_NAME` constant in favor of now deprecated `HighlightingEngine.ACTIVE_CLASS_NAME`
+
+- 6962fcf: Added `maxSize` property to `InputEditorSizeParams` interface of `@itwin/appui-abstract` package (using TypeScript declaration merging) to allow setting the max size for input editors.
+
+### Patch Changes
+
+- @itwin/core-react@5.28.0
+
+## 5.27.1
+
+### Patch Changes
+
+- Updated dependencies [47cf43f]
+  - @itwin/core-react@5.27.1
+
+## 5.27.0
+
+### Patch Changes
+
+- @itwin/core-react@5.27.0
+
 ## 5.26.0
 
 ### Patch Changes